Node/CCQ: Move watcher code
